organize front-end javascripts
