+++ client-js/QueryChartOnly.js
@@ -1,6 +1,6 @@
 var React = require('react')
-var fetchJson = require('./fetch-json.js')
-var SqlpadTauChart = require('./SqlpadTauChart.js')
+var fetchJson = require('./utilities/fetch-json.js')
+var SqlpadTauChart = require('./components/SqlpadTauChart.js')
 var DropdownButton = require('react-bootstrap/lib/DropdownButton')
 var MenuItem = require('react-bootstrap/lib/MenuItem')
 import IncompleteDataNotification from './components/IncompleteDataNotification'
