+++ client-js/components/QueryResultDataTable.js
